Fixing bugs on EdgeHistogram extractor and in Wavelet class (which impacts Haar and Daubechies extractors).
EdgeHistogram problem: When generating the image signature, this method performs a for loop incrementing it by the "blockSize" value. However, it is not guaranteed that this value is different than 0 (depending on image resolution).
Solution: Updated the method "calculateBlockSize()", which determines the "blockSize" value, to return at least 1. Moreover, the number of blocks are adjusted accordingly.

Wavelet class problem: Uninitialized variable "multiSquare", that can give a Segmentation Fault on class destructor, when comparing it with NULL.
Solution: Updating class constructor by assigning NULL to this variable.

Signed-off-by: lscabora <lscabora@user72.morpheus.uspnet.usp.br>